Wrote tests for Response class
Wrote tests covering all specified behaviors of the
`Zend\Feed\Reader\Http\Response` implementation. In doing so, realized
that numeric (integer, float) values are allowed for some headers, so
fixed that behavior. Additionally, this exposed several of the exception
issues that @Maks3w raised in the review, providing reasonable context
for addressing and fixing them.