Made zend-feed forwards compatible with zend-stdlib and zend-servicemanager v3
Updated `composer.json` to pin to stable versions, and, in the case of
zend-stdlib and zend-servicemanager, both v2 and v3 versions.

Updated tests to skip tests that depend on components not yet
forwards-compatible with v3 versions; these conditionals can be removed in the
future when those components are ready. Travis now removes these components when
testing against zend-servicemanager v3.