Updates tests to no longer use HTTP_RAW_POST_DATA
Adds a property to `AbstractCallback`, `$inputStream`, which defaults to
`php://input`, but which allows a string file/stream name, or a
resource.

Updates `AbstractCallback::_getRawBody()` to vary how it reads the body
based on whether `$inputStream` is a resource, using
`stream_get_contents()` if it is, and `file_get_contents()` otherwise.

`CallbackTest` now defines `mockInputStream()`, which accepts an
`AbstractCallback` instance and the contents to inject in the input
stream. It creates a new `php://temp` resource with the `$contents`
provided (rewinding the handle to mimic `php://input`), and then uses
`ReflectionProperty` to change the value of the callback's
`$inputStream` property to point to the new resource.

All tests that previously used `HTTP_RAW_POST_DATA` were updated to
call `mockInputStream()` with the callback instance and contents to
seed.