Adds Writer support for itunes:isClosedCaptioned tag to ITunes extension
Adds the method `Zend\Feed\Writer\Extension\ITunes\Entry::setItunesIsClosedCaptioned(bool $status) : self`.
The method raises an exception for non-boolean `$status` values. If the
value is boolean true, it stores it; otherwise it ignores it.

`Zend\Feed\Writer\Extension\ITunes\Renderer\Entry` now renders the
`itunes:isClosedCaptioned` tag if the entry indicates it is, using the
value `Yes` as the contents.