Allow usage of Reader without GooglePlayPodcast extension
Custom implementations of `ExtensionManagerInterface` may not have _new
core_ extensions present. As a result, when upgrading, an exception is
thrown due to inability to load the new extension.

This was discovered with the 2.10 release, when we added a new core
extension, GooglePlayPodcast; see #80 and https://www.drupal.org/project/drupal/issues/2976335
for more details.

As such, we now check for _new_ core extensions before registering them.
When we discover they are not in the extension manager, we emit an
`E_USER_NOTICE` indicating the user should update their
`ExtensionManagerInterface` implementation to add entries for the new
extension, but then continue on without registering the extension. This
ensures no exception is thrown.